+++ e2e/tests/insertPlugins.e2e.ts
@@ -1,23 +1,18 @@
-/* eslint-disable new-cap */
-/*global cy*/
 import {
   TOOLBARS,
   STATIC_TOOLBAR_BUTTONS_WITHOUT_EMBED,
-  // STATIC_TOOLBAR_BUTTONS_MEDIA,
   VIDEO_PLUGIN,
   SOUND_CLOUD,
   SOCIAL_EMBED,
-  // GIPHY_PLUGIN,
 } from '../cypress/dataHooks';
 import { DEFAULT_DESKTOP_BROWSERS } from './settings';
-import { useUploadConfig } from '../cypress/testAppConfig';
 
 const LINKS = {
   YOUTUBE: 'https://www.youtube.com/watch?v=whbidPR4nVA',
   SOUNDCLOUD: 'https://soundcloud.com/martingarrix/martin-garrix-animals-original',
 };
 
-const modalHandler = (INPUT, ADD_BUTTON, LINK) => {
+const modalHandler = (INPUT: string, ADD_BUTTON: string, LINK: string) => {
   cy.get(`[data-hook=${INPUT}]`)
     .wait(500)
     .type(LINK);
@@ -51,7 +46,7 @@ const additionalCommands = {
   // ADSENSE: () => {},
 };
 
-const testInsertPlugin = toolbar => ([plugin, pluginButtonName]) =>
+const testInsertPlugin = (toolbar: string) => ([plugin, pluginButtonName]: [string, string]) =>
   it(`should insert ${plugin?.toLocaleLowerCase()}`, function() {
     cy.loadRicosEditorAndViewer('empty')
       .wait(500)
@@ -64,36 +59,36 @@ const testInsertPlugin = toolbar => ([plugin, pluginButtonName]) =>
     cy.eyesCheckWindow(this.test.title);
   });
 
-const testNativeUploadMediaPlugin = toolbar => ([plugin, pluginButtonName]) =>
-  it(`should upload native ${plugin?.toLocaleLowerCase()}`, function() {
-    const testAppConfig = {
-      ...useUploadConfig,
-    };
-    cy.loadRicosEditorAndViewer('empty', testAppConfig)
-      .wait(500)
-      .insertPlugin(toolbar, pluginButtonName)
-      .then(el => {
-        const mockFileList = new DataTransfer();
-        const file = new File(['image'], 'native.jpg', { type: 'image/png' });
-        mockFileList.items.add(file);
-        el[0].files = mockFileList.files;
-        el[0].dispatchEvent(
-          new Event('change', {
-            bubbles: true,
-          })
-        );
-      });
+// const testNativeUploadMediaPlugin = toolbar => ([plugin, pluginButtonName]) =>
+//   it(`should upload native ${plugin?.toLocaleLowerCase()}`, function() {
+//     const testAppConfig = {
+//       ...useUploadConfig,
+//     };
+//     cy.loadRicosEditorAndViewer('empty', testAppConfig)
+//       .wait(500)
+//       .insertPlugin(toolbar, pluginButtonName)
+//       .then(el => {
+//         const mockFileList = new DataTransfer();
+//         const file = new File(['image'], 'native.jpg', { type: 'image/png' });
+//         mockFileList.items.add(file);
+//         el[0].files = mockFileList.files;
+//         el[0].dispatchEvent(
+//           new Event('change', {
+//             bubbles: true,
+//           })
+//         );
+//       });
 
-    cy.wait(2000);
+//     cy.wait(2000);
 
-    cy.eyesCheckWindow(this.test.title);
-  });
+//     cy.eyesCheckWindow(this.test.title);
+//   });
 
 const eyesOpen = ({
   test: {
     parent: { title },
   },
-}) =>
+}: Mocha.Context) =>
   cy.eyesOpen({
     appName: 'Insert Plugins',
     testName: title,